button apply and midify, bug fix

+++ GB_moudle/GB_moudles.h
@@ -10,7 +10,7 @@ typedef struct point_info
     LPARAM          lparm;      //좌표  
     int             cWidth;     //굵기
     COLORREF        color;      //색상
-    DWORD           ctime;      //시간
+    ULONGLONG           ctime;      //시간
     UINT            state;      //상태{WM_LBUTTONDOWN }
 }PINFO;
